:recycle: 서브 컴포넌트를 추출하여 메인 컴포넌트 간소화
