+++ src/app/seok/kmong/components/banner.tsx
@@ -1,11 +1,9 @@
 export const Banner = () => {
   return (
-    <div className=" w-full h-12 bg-black flex items-center justify-center">
+    <div className="flex items-center justify-center w-full h-12 bg-black ">
       <div className="text-white">
         <p>
-          신규 가입하면{' '}
-          <span className="text-[#f5c126] font-bold">10만원 쿠폰</span> 드려요
-          &gt;
+          신규 가입하면 <span className="text-[#f5c126] font-bold">10만원 쿠폰</span> 드려요 &gt;
         </p>
       </div>
     </div>
