+++ src/app/seok/kmong/page.tsx
@@ -1,8 +1,16 @@
-import React from "react"
-import { Header } from "./components/header"
+import React from 'react'
+import { Header } from './components/header/header'
+import { Banner } from './components/banner'
+import { Category } from './components/category/category'
 
 const page = () => {
-  return <Header />
+  return (
+    <>
+      <Banner />
+      <Header />
+      <Category />
+    </>
+  )
 }
 
 export default page
